backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
dnspython
bcrypt
